+++ src/main/java/org/accounting/system/services/groupmanagement/OIDCGroupManagementFactory.java
@@ -3,13 +3,9 @@ package org.accounting.system.services.groupmanagement;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import org.accounting.system.beans.RequestUserContext;
-import org.accounting.system.repositories.OidcTenantConfigRepository;
 
 @ApplicationScoped
-public class GroupManagementFactory {
-
-    @Inject
-    OidcTenantConfigRepository oidcTenantConfigRepository;
+public class OIDCGroupManagementFactory implements GroupManagementFactoryI {
 
     @Inject
     RequestUserContext requestUserContext;